refactor: fix migrate command PR comments

+++ lib/controllers/migrate-controller.ts
@@ -1,9 +1,9 @@
 import * as path from "path";
 import * as semver from "semver";
 import * as constants from "../constants";
-import { BaseUpdateController } from "./base-update-controller";
+import { UpdateControllerBase } from "./update-controller-base";
 
-export class MigrateController extends BaseUpdateController implements IMigrateController {
+export class MigrateController extends UpdateControllerBase implements IMigrateController {
 	constructor(
 		protected $fs: IFileSystem,
 		protected $platformCommandHelper: IPlatformCommandHelper,
@@ -12,6 +12,7 @@ export class MigrateController extends BaseUpdateController implements IMigrateC
 		protected $packageManager: IPackageManager,
 		private $devicePlatformsConstants: Mobile.IDevicePlatformsConstants,
 		private $logger: ILogger,
+		private $errors: IErrors,
 		private $addPlatformService: IAddPlatformService,
 		private $pluginsService: IPluginsService,
 		private $projectDataService: IProjectDataService) {
@@ -28,15 +29,15 @@ export class MigrateController extends BaseUpdateController implements IMigrateC
 	];
 
 	static readonly migrationDependencies : IMigrationDependency[] = [
-		{ packageName: constants.TNS_CORE_MODULES_NAME, isDev: false, verifiedVersion: "6.0.0-next-2019-06-10-092158-01"},
-		{ packageName: constants.TNS_CORE_MODULES_WIDGETS_NAME, isDev: false, verifiedVersion: "6.0.0-next-2019-06-10-092158-01"},
+		{ packageName: constants.TNS_CORE_MODULES_NAME, verifiedVersion: "6.0.0-next-2019-06-10-092158-01"},
+		{ packageName: constants.TNS_CORE_MODULES_WIDGETS_NAME, verifiedVersion: "6.0.0-next-2019-06-10-092158-01"},
 		{ packageName: "node-sass", isDev: true, verifiedVersion: "4.12.0"},
 		{ packageName: "typescript", isDev: true, verifiedVersion: "3.4.1"},
 		{ packageName: "less", isDev: true, verifiedVersion: "3.9.0"},
 		{ packageName: "nativescript-dev-sass", isDev: true, replaceWith: "node-sass"},
 		{ packageName: "nativescript-dev-typescript", isDev: true, replaceWith: "typescript"},
 		{ packageName: "nativescript-dev-less", isDev: true, replaceWith: "less"},
-		{ packageName: constants.WEBPACK_PLUGIN_NAME, isDev: true, shouldAdd: true, verifiedVersion: "0.25.0-webpack-2019-06-11-105349-01"},
+		{ packageName: constants.WEBPACK_PLUGIN_NAME, isDev: true, shouldAddIfMissing: true, verifiedVersion: "0.25.0-webpack-2019-06-11-105349-01"},
 		{ packageName: "nativescript-camera", verifiedVersion: "4.5.0"},
 		{ packageName: "nativescript-geolocation", verifiedVersion: "5.1.0"},
 		{ packageName: "nativescript-imagepicker", verifiedVersion: "6.2.0"},
@@ -50,16 +51,16 @@ export class MigrateController extends BaseUpdateController implements IMigrateC
 		{ packageName: "nativescript-ui-autocomplete", verifiedVersion: "5.0.0-androidx-110619"},
 		{ packageName: "nativescript-datetimepicker", verifiedVersion: "1.1.0"},
 		//TODO update with compatible with webpack only hooks
-		{ packageName: "kinvey-nativescript-sdk", verifiedVersion: "4.1.0"},
+		{ packageName: "kinvey-nativescript-sdk", verifiedVersion: "4.2.1"},
 		//TODO update with compatible with webpack only hooks
-		{ packageName: "nativescript-plugin-firebase", verifiedVersion: "8.3.2"},
+		{ packageName: "nativescript-plugin-firebase", verifiedVersion: "9.0.1"},
 		//TODO update with no prerelease version compatible with webpack only hooks
 		{ packageName: "nativescript-vue", verifiedVersion: "2.3.0-rc.0"},
 		{ packageName: "nativescript-permissions", verifiedVersion: "1.3.0"},
 		{ packageName: "nativescript-cardview", verifiedVersion: "3.2.0"}
 	];
 
-	static readonly tempFolder: string = ".migration_backup";
+	static readonly backupFolder: string = ".migration_backup";
 	static readonly migrateFailMessage: string = "Could not migrate the project!";
 	static readonly backupFailMessage: string = "Could not backup project folders!";
 
@@ -72,15 +73,15 @@ export class MigrateController extends BaseUpdateController implements IMigrateC
 
 	public async migrate({projectDir}: {projectDir: string}): Promise<void> {
 		const projectData = this.$projectDataService.getProjectData(projectDir);
-		const tmpDir = path.join(projectDir, MigrateController.tempFolder);
+		const backupDir = path.join(projectDir, MigrateController.backupFolder);
 
 		try {
 			this.$logger.info("Backup project configuration.");
-			this.backup(MigrateController.folders, tmpDir, projectData);
+			this.backup(MigrateController.folders, backupDir, projectData.projectDir);
 			this.$logger.info("Backup project configuration complete.");
 		} catch (error) {
 			this.$logger.error(MigrateController.backupFailMessage);
-			this.$fs.deleteDirectory(tmpDir);
+			this.$fs.deleteDirectory(backupDir);
 			return;
 		}
 
@@ -88,26 +89,34 @@ export class MigrateController extends BaseUpdateController implements IMigrateC
 			await this.cleanUpProject(projectData);
 			await this.migrateDependencies(projectData);
 		} catch (error) {
-			this.restoreBackup(MigrateController.folders, tmpDir, projectData);
+			this.restoreBackup(MigrateController.folders, backupDir, projectData.projectDir);
 			this.$logger.error(MigrateController.migrateFailMessage);
 		}
 	}
 
 	public async shouldMigrate({projectDir}: IProjectDir): Promise<boolean> {
 		const projectData = this.$projectDataService.getProjectData(projectDir);
+
 		for (let i = 0; i < MigrateController.migrationDependencies.length; i++) {
 			const dependency = MigrateController.migrationDependencies[i];
-			const collection = dependency.isDev ? projectData.devDependencies : projectData.dependencies;
-			if (dependency.replaceWith && collection && collection[dependency.packageName]) {
+			const hasDependency = this.hasDependency(dependency, projectData);
+
+			if (hasDependency && dependency.replaceWith) {
 				return true;
 			}
 
-			if (!this.shouldSkipMigrationDependency(dependency, projectData) && await this.shouldMigrateDependencyVersion(dependency, projectData)) {
+			if (hasDependency && await this.shouldMigrateDependencyVersion(dependency, projectData)) {
+				return true;
+			}
+
+			if (!hasDependency && dependency.shouldAddIfMissing) {
 				return true;
 			}
 		}
+
 		for (const platform in this.$devicePlatformsConstants) {
-			if (await this.shouldUpdateRuntimeVersion({ targetVersion: this.verifiedPlatformVersions[platform.toLowerCase()], platform, projectData, shouldAdd: true})) {
+			const hasRuntimeDependency = this.hasRuntimeDependency({platform, projectData});
+			if (!hasRuntimeDependency || await this.shouldUpdateRuntimeVersion({ targetVersion: this.verifiedPlatformVersions[platform.toLowerCase()], platform, projectData})) {
 				return true;
 			}
 		}
@@ -125,27 +134,39 @@ export class MigrateController extends BaseUpdateController implements IMigrateC
 	}
 
 	private async migrateDependencies(projectData: IProjectData): Promise<void> {
-		this.$logger.info("Start migrating dependencies.");
+		this.$logger.info("Start dependencies migration.");
 		for (let i = 0; i < MigrateController.migrationDependencies.length; i++) {
 			const dependency = MigrateController.migrationDependencies[i];
-			if (this.shouldSkipMigrationDependency(dependency, projectData)) {
+			const hasDependency = this.hasDependency(dependency, projectData);
+
+			if (hasDependency && dependency.replaceWith) {
+				this.$pluginsService.removeFromPackageJson(dependency.packageName, dependency.isDev, projectData.projectDir);
+				const replacementDep = _.find(MigrateController.migrationDependencies, migrationPackage => migrationPackage.packageName === dependency.replaceWith);
+				if (!replacementDep) {
+					this.$errors.failWithoutHelp("Failed to find replacement dependency.");
+				}
+				this.$logger.info(`Replacing '${dependency.packageName}' with '${replacementDep.packageName}'.`, );
+				this.$pluginsService.addToPackageJson(replacementDep.packageName, replacementDep.verifiedVersion, replacementDep.isDev, projectData.projectDir);
 				continue;
 			}
 
-			if (dependency.replaceWith) {
-				this.$pluginsService.removeFromPackageJson(dependency.packageName, dependency.isDev, projectData.projectDir);
-				const replacementDep = _.find(MigrateController.migrationDependencies, migrationPackage => migrationPackage.packageName === dependency.replaceWith);
-				this.$logger.info(`Replacing '${dependency.packageName}' with '${replacementDep.packageName}'.`, );
-				this.$pluginsService.addToPackageJson(replacementDep.packageName, replacementDep.verifiedVersion, replacementDep.isDev, projectData.projectDir);
-			} else if (await this.shouldMigrateDependencyVersion(dependency, projectData)) {
+			if (hasDependency && await this.shouldMigrateDependencyVersion(dependency, projectData)) {
 				this.$logger.info(`Updating '${dependency.packageName}' to compatible version '${dependency.verifiedVersion}'`);
 				this.$pluginsService.addToPackageJson(dependency.packageName, dependency.verifiedVersion, dependency.isDev, projectData.projectDir);
+				continue;
+			}
+
+			if (!hasDependency && dependency.shouldAddIfMissing) {
+				this.$logger.info(`Adding '${dependency.packageName}' with version '${dependency.verifiedVersion}'`);
+				this.$pluginsService.addToPackageJson(dependency.packageName, dependency.verifiedVersion, dependency.isDev, projectData.projectDir);
+				continue;
 			}
 		}
 
 		for (const platform in this.$devicePlatformsConstants) {
 			const lowercasePlatform = platform.toLowerCase();
-			if (await this.shouldUpdateRuntimeVersion({targetVersion: this.verifiedPlatformVersions[lowercasePlatform], platform, projectData, shouldAdd: true})) {
+			const hasRuntimeDependency = this.hasRuntimeDependency({platform, projectData});
+			if (!hasRuntimeDependency || await this.shouldUpdateRuntimeVersion({targetVersion: this.verifiedPlatformVersions[lowercasePlatform], platform, projectData})) {
 				const verifiedPlatformVersion = this.verifiedPlatformVersions[lowercasePlatform];
 				const platformData = this.$platformsDataService.getPlatformData(lowercasePlatform, projectData);
 				this.$logger.info(`Updating ${platform} platform to version '${verifiedPlatformVersion}'.`);
@@ -166,38 +187,15 @@ export class MigrateController extends BaseUpdateController implements IMigrateC
 
 	private async shouldMigrateDependencyVersion(dependency: IMigrationDependency, projectData: IProjectData): Promise<boolean> {
 		const collection = dependency.isDev ? projectData.devDependencies : projectData.dependencies;
-		if (
-			collection &&
-			collection[dependency.packageName]
-		) {
-			const maxSatisfyingVersion = await this.getMaxDependencyVersion(dependency.packageName, collection[dependency.packageName]);
-			if (maxSatisfyingVersion) {
-				const isPrereleaseVersion = semver.prerelease(maxSatisfyingVersion);
-				const coerceMaxSatisfying = semver.coerce(maxSatisfyingVersion).version;
-				const coerceVerifiedVersion = semver.coerce(dependency.verifiedVersion).version;
-				//This makes sure that if the user has a prerelease 6.0.0-next-2019-06-10-092158-01 version we will update it to 6.0.0
-				if (isPrereleaseVersion) {
-					if (semver.gt(coerceMaxSatisfying, coerceVerifiedVersion)) {
-						return false;
-					}
+		const maxSatisfyingVersion = await this.getMaxDependencyVersion(dependency.packageName, collection[dependency.packageName]);
 
-					//TODO This should be removed once we update the verified versions to no prerelease versions
-					if (semver.eq(maxSatisfyingVersion, dependency.verifiedVersion)) {
-						return false;
-					}
-				} else if (semver.gte(coerceMaxSatisfying, coerceVerifiedVersion)) {
-					return false;
-				}
-			}
-		}
-
-		return true;
+		return !(maxSatisfyingVersion && semver.gte(maxSatisfyingVersion, dependency.verifiedVersion));
 	}
 
-	private shouldSkipMigrationDependency(dependency: IMigrationDependency, projectData: IProjectData) {
-		if (!dependency.shouldAdd) {
-			return this.shouldSkipDependency(dependency, projectData);
-		}
+	protected async shouldUpdateRuntimeVersion({targetVersion, platform, projectData}: {targetVersion: string, platform: string, projectData: IProjectData}): Promise<boolean> {
+		const maxRuntimeVersion = await this.getMaxRuntimeVersion({platform, projectData});
+
+		return !(maxRuntimeVersion && semver.gte(maxRuntimeVersion, targetVersion));
 	}
 }
